Conflicts:
	lib/manga-squirrel/cli.rb
	lib/manga-squirrel/queuer.rb

+++ lib/manga-squirrel/worker.rb
@@ -4,50 +4,50 @@ require 'open-uri'
 require 'fileutils'
 require 'zip/zip'
 require 'manga-squirrel/common'
+require 'peach'
 
 module Manga
   module Squirrel
     class Manga::Squirrel::Worker
       @queue  = 'manga-squirrel'
 
-      def self.perform(action, options)
-        options = Hash.transform_keys_to_symbols(options)
+      def self.perform(action, chapter)
+        chapter = Hash.transform_keys_to_symbols(chapter)
         case action
         when QueueAction::Download
-          self.doDownload options[:chapter], options[:page], options[:url]
+          self.doDownload chapter
         when QueueAction::Archive
-          self.doArchive options[:root], options[:chapter], options[:outdir]
+          self.doArchive chapter
         end
       end
 
-      def self.doDownload(chapter, page, url)
-        doc = Nokogiri::HTML(open(url))
+      def self.doDownload(chapter)
+        chapter[:pages].peach { 
+        |page|
+          page = Hash.transform_keys_to_symbols(page)
+          doc = Nokogiri::HTML(open(page[:url]))
 
-        img = doc.css("#"+chapter[:img_div]).attribute('src').value
-        ext = img.gsub(/\.*(\.[^\.]*)$/).first
+          img = doc.css(chapter[:img_div]).attribute('src').value
+          ext = img.gsub(/\.*(\.[^\.]*)$/).first
 
-        FileUtils.mkdir_p dir = gendir(chapter)
+          FileUtils.mkdir_p dir = gendir(chapter)
 
-        system 'curl', img, "-o", File.join(dir, "#{"%03d" % page}#{ext}")
+          system 'curl', img, "-o", File.join(dir, "#{"%03d" % page[:num]}#{ext}")
+        }
       end
 
-      def self.doArchive(root, chapter, out)
-        dir = File.join(out, chapter)
+      def self.doArchive(chapter)
+        dir = File.join(File.expand_path("."), chapter)
+        FileUtils.mkdir_p(File.dirname(dir)) unless File.directory?(File.dirname(dir))
 
-        if not File.directory?(File.join(out,File.dirname(chapter))) then
-          FileUtils.mkdir_p File.join(out,File.dirname(chapter))
-        end
-
-        if File.exists?(dir+".cbz") then
-          File.delete(dir+".cbz")
-        end
+        File.delete(dir+".cbz") if File.exists?(dir+".cbz")
 
         Zip::ZipFile.open(dir+".cbz", Zip::ZipFile::CREATE) { 
           |zipfile|
-          puts Dir.entries(File.join(root,chapter)).inspect
-          Dir.entries(File.join(root, chapter)).grep(/^[^.]/).sort.each { 
+          puts Dir.entries(dir).inspect
+          Dir.entries(dir).grep(/^[^.]/).sort.each { 
             |file|
-            zipfile.add(File.basename(file), File.join(root, chapter, file))
+            zipfile.add(File.basename(file), File.join(dir, file))
           }
         }
       end
